Change data type

+++ app/src/main/java/cz/applifting/humansis/model/api/DistributeSmartcardRequest.kt
@@ -1,7 +1,7 @@
 package cz.applifting.humansis.model.api
 
 class DistributeSmartcardRequest (
-    val value: Int = 1,
+    val value: Double = 1.0,
     val distributionId: Int,
     val createdAt: String
 )
