+++ app/src/main/java/cz/applifting/humansis/model/db/CommodityLocal.kt
@@ -2,4 +2,4 @@ package cz.applifting.humansis.model.db
 
 import cz.applifting.humansis.model.CommodityType
 
-data class CommodityLocal(val type: CommodityType, val value: Int, val unit: String)
+data class CommodityLocal(val type: CommodityType, val value: Double, val unit: String)
